backend/requirements.txt
# Django核心
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# 数据库
mysqlclient>=2.2.0
PyMySQL>=1.1.0

# 认证和安全
djangorestframework-simplejwt>=5.3.0
bcrypt>=4.1.0
cryptography>=41.0.0

# 缓存和任务
redis>=5.0.0
celery>=5.3.0
django-celery-beat>=2.5.0

# 文件存储
oss2>=2.18.0

# 工具类
python-decouple>=3.8
Pillow>=10.1.0

# 开发工具
django-debug-toolbar>=4.2.0

# API文档
drf-yasg>=1.21.7

# AI服务
dashscope>=1.17.0  # 阿里万相（通义万相）SDK
